+++ ssdp-rs/src/message/header.rs
@@ -22,7 +22,7 @@ use std::{
 use uuid::Uuid;
 
 use crate::{
-    MULTICAST,
+    MULTICAST, SSDP_PORT,
     message::{DeviceDetails, ParseError, ServiceDetails},
 };
 
@@ -288,6 +288,62 @@ impl Display for ST {
             ST::Uuid(uuid) => write!(f, "uuid:device-{}", uuid),
             ST::Device(device_details) => write!(f, "urn:{device_details}"),
             ST::Service(service_details) => write!(f, "urn:{service_details}"),
+        }
+    }
+}
+
+#[derive(Debug, Default, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
+/// The port specified in a Upnp message.
+///
+/// Treat this like a semantically specific `Option<u16>` with a valuable implementation of
+/// `Default`, `From<Option<u16>>` & `Into<u16>`.
+pub enum UpnpPort {
+    /// Specifically defined value.
+    ///
+    /// If this is set to the default [SSDP_PORT], then it means the message specifically
+    /// defined that value.
+    Defined(u16),
+    /// A value was not defined. Conversion to a `u16` will provide the default [SSDP_PORT]
+    #[default]
+    Default,
+}
+
+impl Header for UpnpPort {
+    const HEADER_KEY: &'static str = "SEARCHPORT.UPNP.ORG";
+}
+
+/// `None` maps to `Default`
+impl TryFrom<Option<&str>> for UpnpPort {
+    type Error = ParseError;
+
+    fn try_from(port: Option<&str>) -> Result<Self, Self::Error> {
+        Ok(port
+            .map(|port| {
+                port.parse::<u16>()
+                    .map_err(|_| ParseError::InvalidPort(port.to_string()))
+            })
+            .transpose()?
+            .into())
+    }
+}
+/// `None` maps to `Default`
+impl From<Option<u16>> for UpnpPort {
+    fn from(port: Option<u16>) -> Self {
+        match port {
+            Some(port) => Self::Defined(port),
+            None => Self::Default,
+        }
+    }
+}
+
+/// Provides:
+/// - `Defined(port)`: the defined port
+/// - `Default`: [SSDP_PORT]
+impl From<UpnpPort> for u16 {
+    fn from(port: UpnpPort) -> Self {
+        match port {
+            UpnpPort::Defined(p) => p,
+            UpnpPort::Default => SSDP_PORT,
         }
     }
 }
